+++ benchmark/readme_validation.py
@@ -18,7 +18,7 @@ from datetime import datetime
 import matplotlib.pyplot as plt
 import numpy as np
 
-import geopack
+from mageometry import geopack
 
 # ---------------------------------------------------------------------------
 # Parameters